+++ drake/systems/plants/parser_sdf.cc
@@ -501,7 +501,7 @@ void ParseSdfJoint(RigidBodyTree* model, std::string model_name,
   Isometry3d transform_to_parent_body =
       transform_parent_to_model.inverse() * transform_to_model;
 
-  if (child->hasParent()) {
+  if (child->has_mobilizer_joint()) {
     // ... then implement it as a loop joint.
 
     // Gets the loop point in the joint's reference frame. Since the SDF